requirements.txt
torch
torchvision
numpy
pandas
matplotlib
seaborn
Pillow
scikit-learn

Install them using: pip install -r requirements.txt